requirements.txt
requests
websocket_client
./dist/NorenRestApiPy-0.0.20-py2.py3-none-any.whl
pandas
pyyaml
tkinter
threading

shoonya broker credentials/api key